# sdk/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST v1 with decimal-string IDs and timestamps, scoped bearer authentication and bounded requests. Input schemas are shared with MCP. Responses remain extensible JSON objects; this specification is not a complete generated response type library.",
    "title": "ChronoDB Community API",
    "version": "0.4.0-alpha.3"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/add_edges": {
      "post": {
        "description": "Compatibility alias for ingest_edges with the same durability and scope rules.",
        "operationId": "add_edges",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "durability": {
                    "description": "Omitted uses schema settings. fsync acknowledges disk synchronization; buffered needs a later sync",
                    "enum": [
                      "buffered",
                      "fsync"
                    ],
                    "type": "string"
                  },
                  "edges": {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "dst": {
                          "description": "Exact u64 ID as a decimal string",
                          "pattern": "^[0-9]+$",
                          "type": "string"
                        },
                        "kind": {
                          "maximum": 65535,
                          "minimum": 0,
                          "type": "integer"
                        },
                        "payload": {
                          "pattern": "^[0-9a-fA-F]{32}$",
                          "type": "string"
                        },
                        "properties": {
                          "description": "Values matching schema property names/types. Use this or payload. u64/i64 values are decimal strings.",
                          "type": "object"
                        },
                        "src": {
                          "description": "Exact u64 ID as a decimal string",
                          "pattern": "^[0-9]+$",
                          "type": "string"
                        },
                        "valid_from": {
                          "description": "Exact i64 microseconds as a decimal string",
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "valid_to": {
                          "description": "Exact i64 microseconds as a decimal string",
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "src",
                        "dst",
                        "kind",
                        "valid_from"
                      ],
                      "type": "object"
                    },
                    "maxItems": 10000,
                    "minItems": 1,
                    "type": "array"
                  },
                  "fork": {
                    "description": "Optional selected fork ID; omitted uses the parent. Edge IDs belong to the selected fork. Query/write times must be at or after the fork time.",
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  }
                },
                "required": [
                  "edges"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Success. See API.md for operation-specific response fields."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Structured API error; 429/503 may include Retry-After. Proxies may return non-JSON bodies.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-required-scope": "ingest"
      }
    },
    "/v1/add_node": {
      "post": {
        "description": "Register an isolated node. Ingest/admin scope; repeating it is a no-op.",
        "operationId": "add_node",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "durability": {
                    "description": "Omitted uses schema settings. fsync acknowledges disk synchronization; buffered needs a later sync",
                    "enum": [
                      "buffered",
                      "fsync"
                    ],
                    "type": "string"
                  },
                  "fork": {
                    "description": "Optional selected fork ID; omitted uses the parent. Edge IDs belong to the selected fork. Query/write times must be at or after the fork time.",
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  },
                  "id": {
                    "description": "Exact u64 ID as a decimal string",
                    "pattern": "^[0-9]+$",
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Success. See API.md for operation-specific response fields."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Structured API error; 429/503 may include Retry-After. Proxies may return non-JSON bodies.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-required-scope": "ingest"
      }
    },
...
    "/v1/asset_compose": {
      "post": {
        "description": "Publish a typed asset from 1–16 immutable chunk_v1 uploads in order (≤16 MiB). Safe to retry.",
        "operationId": "asset_compose",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "chunks": {
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 16,
                    "minItems": 1,
                    "type": "array"
                  },
                  "metadata": {
                    "additionalProperties": false,
                    "description": "Tensors require raw_le encoding, non-null dtype and exact row-major little-endian bytes. Scalars have shape []. Non-tensors require null/omitted dtype and empty/omitted shape. 1 byte to 16 MiB per composed asset.",
                    "properties": {
                      "dtype": {
                        "enum": [
                          null,
                          "u8",
                          "i8",
                          "u16",
                          "u32",
                          "u64",
                          "i16",
                          "i32",
                          "i64",
                          "f16",
                          "bf16",
                          "f32",
                          "f64",
                          "bool"
                        ]
                      },
                      "encoding": {
                        "maxLength": 80,
                        "minLength": 1,
                        "type": "string"
                      },
                      "kind": {
                        "enum": [
                          "tensor",
                          "opaque",
                          "image",
                          "audio",
                          "video"
                        ]
                      },
                      "provenance": {
                        "additionalProperties": {
                          "maxLength": 1024,
                          "type": "string"
                        },
                        "maxProperties": 32,
                        "type": "object"
                      },
                      "shape": {
                        "items": {
                          "minimum": 1,
                          "type": "integer"
                        },
                        "maxItems": 8,
                        "type": "array"
                      },
                      "version": {
                        "const": 1,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "version",
                      "kind",
                      "encoding"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "metadata",
                  "chunks"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Success. See API.md for operation-specific response fields."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Structured API error; 429/503 may include Retry-After. Proxies may return non-JSON bodies.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-required-scope": "ingest"
      }
    },
...
    "/v1/connector_catalog": {
      "post": {
        "description": "Read versioned connector families, presets and exact native/normalized capabilities.",
        "operationId": "connector_catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {},
                "required": [],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "Success. See API.md for operation-specific response fields."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Structured API error; 429/503 may include Retry-After. Proxies may return non-JSON bodies.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-required-scope": "read"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Local development; use HTTPS remotely",
      "url": "http://127.0.0.1:8080"
    }
  ],
...
}